terminado ejercicio 8

+++ tp2/src/utils.py
@@ -9,18 +9,21 @@ def windows[T](arr: list[T], w: int) -> Generator[list[T]]:
         yield arr[i : i + w]
 
 
-def truth_table(n: int, f: Callable[[list[int]], int]) -> tuple[np.ndarray, np.ndarray]:
+def truth_table(
+    n: int,
+    f: Callable[[list[float]], float],
+) -> tuple[np.ndarray, np.ndarray]:
     def bt(xs, ys, acc):
         if len(acc) == n:
             xs.append(np.array(acc))
             ys.append(np.array([f(acc)]))
             return xs, ys
 
-        acc.append(1)
+        acc.append(1.0)
         bt(xs, ys, acc)
         acc.pop()
 
-        acc.append(-1)
+        acc.append(-1.0)
         bt(xs, ys, acc)
         acc.pop()
